r2.14.0

+++ library/dash/src/main/java/com/google/android/exoplayer2/source/dash/offline/DashDownloader.java
@@ -183,7 +183,7 @@ public final class DashDownloader extends SegmentDownloader<DashManifest> {
         continue;
       }
 
-      int segmentCount = index.getSegmentCount(periodDurationUs);
+      long segmentCount = index.getSegmentCount(periodDurationUs);
       if (segmentCount == DashSegmentIndex.INDEX_UNBOUNDED) {
         throw new DownloadException("Unbounded segment index");
       }
